improved file search
now search not only in the description but also in the first filename.

you can combine terms that need to match with spaces.
like: split php
which will fine any gist that has split and php, either in the
description or the filename

and you can also use quotes for exact matches
like "split php"