FEATURE (gist polling): Add gist polling to update changed gists
Implemented a polling mechanism that runs every 10 minutes or whenever
the update button at the top header is pressed.

Once the update starts the update indicator will start spinning until
polling will finish processing all paginated results.

If the currently viewed gist was updated the gist will not update in
favor of keeping changes the user may have made.

If there were no changes to the gist and it was navigated from in the
next polling cycle the gist will be updated.
If there were changes to the gist it will keep the current changes are
these are considered the latest changes and are newer than the polled
update that was attempted.

In the future it is possible to raise a notification to the user that
the gist has attempted to update but failed to due to the gist being
open/being edited and allow the user to choose which version of the gist
he would like to keep.
As well as adding an option in the settings to change the duration
between each polling request, but must take into consideration rate
limits on GitHub.

Closes #124