FIX (toggle directive): Fix toggle directive causing a screen overflow
Fix login triggering enterprise toggle back to false when clicking on it
for the first time.

The overflow was caused by a small div from the toggle css class which
was in absolute position, added back the wrapping toggle element instead
of replacing it and added a position:relative style to the toggle
wrapping tag.

Login screen issue was due to acting twice on the same action, once with
ng-click and once with a watch triggering on the ngmodel.

Closes #125